# backend/docs/swagger.yaml
definitions:
...
  entities.OutBlog:
    properties:
      content:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      pined:
        type: boolean
      slug:
        type: string
      tags:
        items:
          $ref: '#/definitions/entities.Tag'
        type: array
      title:
        type: string
      topics:
        items:
          $ref: '#/definitions/entities.Topic'
        type: array
      updated_at:
        type: string
      visible:
        type: boolean
    type: object
  entities.ReqInBlog:
    properties:
      content:
        type: string
      description:
        type: string
      pined:
        type: boolean
      tags:
        items:
          type: integer
        type: array
      title:
        type: string
      topics:
        items:
          type: integer
        type: array
      visible:
        type: boolean
    type: object
  entities.RetFailed:
    properties:
      error:
        type: string
      msg:
        type: string
      status:
        type: integer
    type: object
  entities.RetSuccess-array_entities_OutBlog:
    properties:
      error:
        type: string
      msg:
        items:
          $ref: '#/definitions/entities.OutBlog'
        type: array
      status:
        type: integer
    type: object
...
  entities.RetSuccess-entities_OutBlog:
    properties:
      error:
        type: string
      msg:
        $ref: '#/definitions/entities.OutBlog'
      status:
        type: integer
    type: object
  entities.RetSuccess-entities_RowsAffected:
    properties:
      error:
        type: string
      msg:
        $ref: '#/definitions/entities.RowsAffected'
      status:
        type: integer
    type: object
...
  entities.RowsAffected:
    properties:
      affectedRows:
        type: integer
    type: object
  entities.Tag:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
  entities.Topic:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: A place to document what I've learned.
  title: Coding Notes
  version: "1.0"
paths:
  /blogs:
    get:
      consumes:
      - application/json
      description: list blogs
      parameters:
      - default: false
        description: show all blogs regardless of visibility or soft delete status
        in: query
        name: all
        type: boolean
      - description: 'filter by topic ids, can be multiple ids. ex: ?topic=1&topic=2'
        in: query
        name: topic
        type: integer
      - description: 'filter by tag ids, can be multiple ids, must be use with topic.
          ex: ?tag=1&tag=2'
        in: query
        name: tag
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RetSuccess-array_entities_OutBlog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.RetFailed'
      summary: List blogs
      tags:
      - blogs
    post:
      consumes:
      - application/json
      description: blogs must have unique titles
      parameters:
      - description: new blog contents
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/entities.ReqInBlog'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RetSuccess-entities_OutBlog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.RetFailed'
      summary: Create blog
      tags:
      - blogs
  /blogs/{id}:
    delete:
      consumes:
      - application/json
      description: update blog
      parameters:
      - description: target blog id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RetSuccess-entities_RowsAffected'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.RetFailed'
      summary: Soft delete blog
      tags:
      - blogs
    get:
      consumes:
      - application/json
      description: get blog
      parameters:
      - description: target blog id
        in: path
        name: id
        required: true
        type: integer
      - default: false
        description: show all blogs regardless of visibility or soft delete status
        in: query
        name: all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RetSuccess-entities_OutBlog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.RetFailed'
      summary: Get blog
      tags:
      - blogs
    put:
      consumes:
      - application/json
      description: update blog
      parameters:
      - description: target blog id
        in: path
        name: id
        required: true
        type: integer
      - description: new blog content
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/entities.ReqInBlog'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RetSuccess-entities_OutBlog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.RetFailed'
      summary: Update blog
      tags:
      - blogs
  /blogs/deleted/{id}:
    delete:
      consumes:
      - application/json
      description: delete blog
      parameters:
      - description: target blog id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RetSuccess-entities_OutBlog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.RetFailed'
      summary: Delete blog
      tags:
      - blogs
    patch:
      consumes:
      - application/json
      description: restore delete blog
      parameters:
      - description: target blog id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RetSuccess-entities_OutBlog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.RetFailed'
      summary: Restore delete blog
      tags:
      - blogs
...
swagger: "2.0"